This patch contains 3 basic modifications necessary for the swekey plugin.
1) Added a PostBody Render Entry Point (conterpart of the prebody renderer)

2) Added the VerifyLogin Entry Point (for both default and LDAP login)
This entry point gives a chance to a plugin to deny a login even if the password is good.
We use it to validate that the swekey is plugged, It can be reused for other plugins (an IP address black lister for example).

3) Added the possibility to add an error message in the login form. So the user will know why the access was denied.
This feature does not work yet because of a bug I found and reported (http://forums.cerb4.com/showthread.php?t=1850).
As soon as your bug will be fixed the feature should work.

For any question you can contact me at luc.andre@musbe.com